renamed "extensions" domain to "extensionManager" for clarity of purpose (per @njx review comment)
